refactor: change names, create types.ts file
