+++ src/components/posts/types.ts
@@ -1,0 +1,8 @@
+export interface IPost {
+  id: number
+  content: string
+}
+export interface IComment {
+  id: number
+  content: string
+}
